requirements.txt
dapr-dev>=1.14.0rc1.dev3
flask-dapr==1.12.0
flask>=3.0.3
pusher>=3.3.2
jinja2>=3.1.4
aiofiles>=24.1.0